Added create actions for entities.
